scripts/translator/requirements.txt
openai
deep_translator
tqdm
google-generativeai
ollama